% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing.R
\name{Bo.Standing}
\alias{Bo.Standing}
\title{Formation Volume Factor by Standing}
\usage{
Bo.Standing(oil.Rsi, gas.sg, oil.api, fluid.tempF)
}
\arguments{
\item{oil.Rsi}{solution oil gas ratio}

\item{gas.sg}{gas specific gravity}

\item{oil.api}{oil specific gravity in API degrees}

\item{fluid.tempF}{fluid temperature in Farenheit}
}
\description{
Bo Standing Correlation
}
