% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pb_coefficients}
\alias{get_pb_coefficients}
\title{Coefficients for Bubble Point Pressure}
\usage{
get_pb_coefficients(correlation)
}
\arguments{
\item{correlation}{Any of these three options: pb, bo, co}
}
\description{
Get Coefficients for Bubble Point Pressure
}
