% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing.R
\name{Rs.Standing}
\alias{Rs.Standing}
\title{Gas Solubility by Standing}
\usage{
Rs.Standing(pres.bubble, gas.sg, oil.api, fluid.tempF)
}
\arguments{
\item{pres.bubble}{Bubble Point Pressure, psi}

\item{gas.sg}{gas specific gravity}

\item{oil.api}{oil specific gravity in API degrees}

\item{fluid.tempF}{fluid temperature in Farenheit}
}
\description{
Rs Standing Correlation
}
