% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_correlations}
\alias{show_correlations}
\title{Show Correlations}
\usage{
show_correlations(table_type)
}
\arguments{
\item{table_type}{any of these types: pb, bo, co}
}
\description{
Show correlation by specifying the type of table.
}
