% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standing.R
\name{Pob.Standing}
\alias{Pob.Standing}
\title{Bubble Point Pressure by Standing}
\usage{
Pob.Standing(oil.Rsi, gas.sg, oil.API, fluid.tempF)
}
\arguments{
\item{oil.Rsi}{solution oil gas ratio}

\item{gas.sg}{gas specific gravity}

\item{oil.API}{oil specific gravity in API degrees}

\item{fluid.tempF}{fluid temperature in Farenheit}
}
\description{
Pb Standing Correlation
}
